+++ src/context/block_unit.rs
@@ -1768,7 +1768,7 @@ impl ContextWriter<'_> {
     eob: u16, pred_mode: PredictionMode, tx_size: TxSize, tx_type: TxType,
     plane_bsize: BlockSize, xdec: usize, ydec: usize,
     use_reduced_tx_set: bool, frame_clipped_txw: usize,
-    frame_clipped_txh: usize, cul_lvl: u8, hash: u64, marker: u16,
+    frame_clipped_txh: usize, cul_lvl: u8, hash: u32, marker: u16,
   ) -> (bool, u8) {
     debug_assert!(frame_clipped_txw != 0);
     debug_assert!(frame_clipped_txh != 0);
